Consolidate markdown path validation logic
