+++ packages/utils/src/fileUtils.ts
@@ -3,12 +3,6 @@ import path from 'path'
 
 import { notFound } from 'next/navigation'
 
-import {
-  WORKSPACE_ROOT,
-  CONTENT_ROOT,
-  MARKDOWN_EXTENSION,
-} from '@/constants/paths'
-
 export function extractSlugFromFilename(filename: string) {
   return path.parse(filename).name
 }
@@ -20,13 +14,6 @@ export function getFilePath(directoryPath: string, slug: string) {
 export function handleFileNotFound(filePath: string) {
   console.error(`File not found: ${filePath}`)
   notFound()
-}
-
-export function isValidMarkdownPath(path: string) {
-  return (
-    path.startsWith(`${WORKSPACE_ROOT}/${CONTENT_ROOT}`) &&
-    path.endsWith(MARKDOWN_EXTENSION)
-  )
 }
 
 export function readFileContents(filePath: string) {
